Prod

+++ GCCBuild/RunWrapper.cs
@@ -36,39 +36,41 @@ namespace GCCBuild
         private readonly ProcessStartInfo startInfo;
         private ShellAppConversion shellApp;
         private string realCommandLine;
+        private string responsefilename = "";
 
         public void Dispose()
         {
             try
             {
-                if (!String.IsNullOrEmpty(shellApp.prerunapp) && startInfo.FileName.Contains("GCCBuildPreRun_"))
+                if (!String.IsNullOrEmpty(shellApp.prerunapp) && !String.IsNullOrEmpty(startInfo.FileName) && startInfo.FileName.Contains("GCCBuildPreRun_"))
                     File.Delete(startInfo.FileName);
 
-                if (!startInfo.Arguments.StartsWith("@") && Path.GetExtension(startInfo.Arguments.Substring(1)) == ".rsp" && File.Exists(startInfo.Arguments.Substring(1)))
-                    File.Delete(startInfo.Arguments.Substring(1));
+                if (!String.IsNullOrEmpty(responsefilename) && File.Exists(responsefilename))
+                    File.Delete(responsefilename);
             }
             catch (Exception ex)
             {
-                Logger.Instance.LogError("~RunWrapper caused an exception:" + ex, shellApp);
+                Logger.Instance.LogError($"~RunWrapper caused an exception shellApp.prerunapp({shellApp.prerunapp}) startInfo.FileName({startInfo.FileName}) responsefilename({responsefilename}):" + ex, shellApp);
             }
         }
 
-        internal RunWrapper(string path, string options, ShellAppConversion shellApp, bool useresponse)
+        internal RunWrapper(string path, string options, ShellAppConversion shellApp, bool supportresponsefile)
         {
-            realCommandLine = $"{path} {options}";
+#if DEBUG
+            Logger.Instance.LogMessage($"RunWrapper({path},{options},{shellApp.shellapp},{supportresponsefile})");
+#endif
+           realCommandLine = $"{path} {options}";
 
-            if (!Utilities.isLinux() && useresponse && (path.Length + options.Length) > 8100) //technically it is 8191
+            if (!Utilities.isLinux() && supportresponsefile && (path.Length + options.Length) > 8100) //technically it is 8191
             {
-                var responsefilename = Path.Combine(shellApp.tmpfolder, "response_" + Guid.NewGuid().ToString() + ".rsp");
+                responsefilename = Path.Combine(shellApp.tmpfolder, "response_" + Guid.NewGuid().ToString() + ".rsp");
                 File.WriteAllText(responsefilename, $"{options}");
                 options = $"@{responsefilename}";
             }
 
             if (!string.IsNullOrEmpty(shellApp.shellapp)) //try to find full path of it from path enviroment!
             {
-                var enviromentPath = System.Environment.GetEnvironmentVariable("PATH");
-                if (!Utilities.isLinux())
-                    enviromentPath = enviromentPath + ";" + Environment.GetEnvironmentVariable("SystemRoot") + @"\sysnative";
+                var enviromentPath = Utilities.GetEnviromentPath();
 
                 //Console.WriteLine(enviromentPath);
                 var paths = enviromentPath.Split(Utilities.isLinux() ? ':' : ';');
@@ -107,7 +109,8 @@ namespace GCCBuild
             }
 
             startInfo = new ProcessStartInfo(path, options);
-			startInfo.UseShellExecute = false;
+			//if you set it to true you can not use redirects and such!
+            startInfo.UseShellExecute = false;
 			startInfo.RedirectStandardError = true;
 			//startInfo.RedirectStandardInput = true;
 			startInfo.RedirectStandardOutput = true;
@@ -173,6 +176,11 @@ namespace GCCBuild
 
             }
             process.WaitForExit();
+            while (!process.HasExited)
+            {
+                Thread.Sleep(50);
+            }
+
             var successfulExit = (process.ExitCode == 0);
 
             if (!String.IsNullOrEmpty(prevErrorRecieved))
@@ -206,6 +214,11 @@ namespace GCCBuild
             ot.Start();
 
             process.WaitForExit();
+            while (!process.HasExited)
+            {
+                Thread.Sleep(50);
+            }
+
             et.Join();
             string output = cv_error;// process.StandardError.ReadToEnd();
             string prevErrorRecieved = "";
@@ -277,6 +290,11 @@ namespace GCCBuild
                 ot.Start();
 
                 process.WaitForExit();
+                while (!process.HasExited)
+                {
+                    Thread.Sleep(50);
+                }
+
                 et.Join();
                 ot.Join();
                 output = /*cv_error +*/ cv_out;
